nFAPI/PNF: Correctly set message length when segmenting
Make the same change as in parent commit, i.e., write the full 32-bit
segment size.

I could not test this, as we do not reach these rx_data.indication
length. It's based on the fix in the parent commit, and to avoid future
bad surprises.

+++ nfapi/open-nFAPI/pnf/public_inc/nfapi_pnf_interface.h
@@ -629,8 +629,8 @@ typedef struct nfapi_pnf_p7_config
 	uint8_t checksum_enabled;
 
 	/*! The maxium size of a P7 segement. If a message is large that this it
-	 * will be segemented */
-	uint16_t segment_size;
+	 * will be segmented. Note: u32 to cover 4G and 5G */
+	uint32_t segment_size;
 
 	/*! The dummy subframe buffer structure that should be used in case there
 	 * are no 'valid' subframe messages */
